Feat: 내부 커맨드 실행 기능 구현 (#30)
1. 내부 커맨드 실행을 위한 exec_another 함수 구현
2. fork 하여 excute 작업을 하는 frist_process 함수 구현

resolved #29

+++ srcs/parsing_utils.c
@@ -24,7 +24,7 @@ int	get_index_redir(char *temp, int index)
 			break ;
 		i++;
 	}
-	if (i == ft_strlen(temp))
+	if (i == (int)ft_strlen(temp))
 		return (-1);
 	return (i);
 }
@@ -61,7 +61,7 @@ char	*sub_make_cmd(char *cmd, char *content, int *i, int *j)
 	char	*new;
 
 	str = cmd;
-	while (*j < ft_strlen(content))
+	while (*j < (int)ft_strlen(content))
 	{
 		*i = get_index_redir(content, *j);
 		if (*i < 0)
